Fixed Issue with Models and Components
Updates to Model objects were not invalidating caches, so when an empty request to Model.prototype.get was made, stale data was being returned.

The data-import function in the page was not updated to handle wurl based query imports into Models, thus the Models are not changing and forcing an update on the component when a page is loaded. The fix is to allow an automatic load of a Model's data when the component is loaded. Also corrected the syntax for the data-import function.