Changes to Model, Getter, Case, Form
Model: Schemas for models are not required to have an "identifier" entry
now. Unique Id's are only necessary to retrieve data from containers.

Added a "getDataStatus" Method to Model Class. This will allow consumers
of the Model's data to check and see if the stored data is determined to
be valid against the Model's schema.

Case: Update to Case to handle changes done to Model.

Getter: Allow the "get" function to have a "stored_data" object passed
along side the query. This can store info that can be later accessed by
the Getter's return functions.

ArrayContainer: Update to ArrayContainer to handle changes done to
Model.