Changed Model setup to use Object.defineProperty and Proxy to define values.
Now a Model's properties can be accessed directly instead of calling .add() or .get()
or accessing .data. Properties will be automatically parsed using the the models schema.
If no schema is provided a Model's constructor, then an AnyModel is returned, which uses a Proxy object to handle updates on the model.

Updated ModelContainer(s), Case, CaseTemplate to handle the new form of Models.

Added toJSON and toJson methods to Models and ModelContainers.

Added Scheduler to keep objects from spamming RAF or setTimeout. Call Scheduler.queueUpdate({object}) to have the object's update method.

TransformTo has been changed to use the Scheduler. Model will also use this when one of its properties is updated.