Rewrote CaseConstructor into a custom HTML parser.
Also rewrote or started to much of the Case stack, including Rivet, Case, CaseTemplate, CaseSkeleton

Created new elements Tap, Pipe, and IO to allow for a more intuitive interface.

	A Tap represents a single property channel that is used to pull and push data from a Model attached to a Case.

	A Pipe takes property values and transforms them into more useful forms.

	An IO represents the tangible connection to DOM, as it maintains it's own element and pushes updates
	from Pipes and sends user input back through the pipes to the Model.

Attribute modification will be possible soon, using essentially the same interface that is set up for element manipulation.

Created a test suite for the CaseConstructor. It still needs to be developed to handle more use cases.

This is a version update to 0.1.0, since the DOM interface is largely changed and many other changes will be made based on the updates of this commit.