Refactored source directory and files.
Got rid of Controller, Element (Renamed element to Component and actually got rid of the Component file.)

Renamed Case to Source and likewise for all derivatives. Case is hard to naturally reason about (it's a case, like a suitecase, that holds and gives access to a Model). Named Source, it's now easier to say "it is the source of access to a Model".

Renamed Linker to Router for easier recognition of the class's usage.
Changed wick.light to wick.startRouting

Reduced several names of different files, such turning model_base.js to base.js (source/model/base.js) and the same for case_base, now (source/source/base.js).

Several other files have been moved as well.

Created an interface for wick, placing all Constructors in wick.core. See the end of wick.js for the full layout. There are also:
wick.model - access to model constructors and factory functions.
wick.startRouting - starts the main wick event loop.
wick.schema - access to schema instances and constructors.
wick.any - factory function for the AnyModel

This is a breaking commit, so the version is now 0.2.0a (appending alpha version tag now)