More file and folder refactoring.
Added attribute handling to the SourceConstructor process. Attributes can be directly affected by adding a Tap with an AttribIO at the end.

Moved common in to the common folder.
Moved view into a new view folder.
wick.js is the only file at the root of ./source.

Added rollup-plugin-gzip for compressing production output.