Updated btree, ScopeContainer, HTMLtags for scope/container
Tags <scope> and <container> are now available to define their respective elements

Adjusted settings for btree to make sure linked containers that retrieve all the contents of a btree container actually work.

Made adjustments to ScopeContainer definition to allow static element trees to define container scopes as apposed to an array of models. Need to create tests for this.