Add selective component rendering within container elements
**useif** binding attributes on children of container elements can be
used to determine what component is selected when mapping component data
models to components. If the evaluation of the expression in the **useif**
attribute binding yields a truthy value, then the associated component
will be used to render the model data.

Component useif binding attributes are evaluated top-down, first match
order. The last component will be used if none of the useif expressions
yield a truthy value.